bug：支持将分区库里的项目的数据迁移至指定数据库问题修复 #9464

+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/service/process/ProcessDataDeleteService.kt
@@ -129,6 +129,8 @@ class ProcessDataDeleteService @Autowired constructor(
                 processDataDeleteDao.deletePipelineWebhookBuildLogDetail(dslContext, projectId, pipelineId)
                 offset += DEFAULT_PAGE_SIZE
             } while (historyInfoRecords?.size == DEFAULT_PAGE_SIZE)
+        }
+        if (!pipelineIds.isNullOrEmpty()) {
             processDataDeleteDao.deletePipelineBuildSummary(dslContext, projectId, pipelineIds)
             processDataDeleteDao.deletePipelineInfo(dslContext, projectId, pipelineIds)
             processDataDeleteDao.deletePipelineLabelPipeline(dslContext, projectId, pipelineIds)
@@ -138,8 +140,8 @@ class ProcessDataDeleteService @Autowired constructor(
             processDataDeleteDao.deletePipelineSetting(dslContext, projectId, pipelineIds)
             processDataDeleteDao.deletePipelineSettingVersion(dslContext, projectId, pipelineIds)
             processDataDeleteDao.deletePipelineBuildHistory(dslContext, projectId, pipelineIds)
-            logger.info("project[$projectId] deleteProjectPipelineRelData success!")
         }
+        logger.info("project[$projectId] deleteProjectPipelineRelData success!")
     }
 
     /**
